requirements.txt
Flask
onvif-zeep
opencv-python-headless
python-dotenv
numpy
